@glwagner 2021-11-17
> Nice one. Should we benchmark anything just to make sure this hasn't affected performance?

That's a good idea!