@simone-silvestri 2021-11-18
I agree very much with this decision, it would be very convenient (and intuitive) to be able to change boundary conditions like that.