@glwagner 2021-12-01
# Benchmarks

## `main`

```
                                            Nonhydrostatic model benchmarks
┌───────────────┬─────────────┬─────┬────────────┬────────────┬────────────┬────────────┬──────────┬────────┬─────────┐
│ Architectures │ Float_types │  Ns │        min │     median │       mean │        max │   memory │ allocs │ samples │
├───────────────┼─────────────┼─────┼────────────┼────────────┼────────────┼────────────┼──────────┼────────┼─────────┤
│           CPU │     Float32 │  32 │   7.865 ms │   8.157 ms │   8.384 ms │   9.923 ms │ 2.50 MiB │   7004 │      10 │
│           CPU │     Float32 │  64 │  13.237 ms │  13.739 ms │  14.162 ms │  16.642 ms │ 3.20 MiB │  12619 │      10 │
│           CPU │     Float32 │ 128 │  49.644 ms │  53.133 ms │  52.873 ms │  56.953 ms │ 3.51 MiB │  16344 │      10 │
│           CPU │     Float32 │ 256 │ 338.545 ms │ 353.497 ms │ 353.048 ms │ 370.172 ms │ 3.81 MiB │  35156 │      10 │
│           CPU │     Float64 │  32 │   9.683 ms │  10.028 ms │  10.324 ms │  12.479 ms │ 3.25 MiB │   7003 │      10 │
│           CPU │     Float64 │  64 │  14.979 ms │  16.581 ms │  17.035 ms │  22.013 ms │ 4.07 MiB │  12652 │      10 │
│           CPU │     Float64 │ 128 │  67.659 ms │  68.736 ms │  70.121 ms │  84.622 ms │ 4.50 MiB │  20205 │      10 │
│           CPU │     Float64 │ 256 │ 513.071 ms │ 520.511 ms │ 520.758 ms │ 531.669 ms │ 5.70 MiB │  85616 │      10 │
│           GPU │     Float32 │  32 │   2.276 ms │   2.324 ms │   2.420 ms │   3.260 ms │ 1.49 MiB │   4670 │      10 │
│           GPU │     Float32 │  64 │   2.260 ms │   2.342 ms │   2.445 ms │   3.493 ms │ 1.48 MiB │   4681 │      10 │
│           GPU │     Float32 │ 128 │   3.319 ms │   3.408 ms │   3.652 ms │   5.730 ms │ 1.48 MiB │   4681 │      10 │
│           GPU │     Float32 │ 256 │  18.567 ms │  19.158 ms │  19.126 ms │  19.527 ms │ 1.48 MiB │   4681 │      10 │
│           GPU │     Float64 │  32 │   2.393 ms │   2.506 ms │   2.631 ms │   3.886 ms │ 1.97 MiB │   4644 │      10 │
│           GPU │     Float64 │  64 │   3.020 ms │   3.154 ms │   3.441 ms │   5.035 ms │ 1.95 MiB │   4655 │      10 │
│           GPU │     Float64 │ 128 │   4.703 ms │   4.853 ms │   5.078 ms │   7.258 ms │ 1.95 MiB │   4655 │      10 │
│           GPU │     Float64 │ 256 │  26.831 ms │  36.351 ms │  35.469 ms │  37.037 ms │ 1.95 MiB │   4655 │      10 │
└───────────────┴─────────────┴─────┴────────────┴────────────┴────────────┴────────────┴──────────┴────────┴─────────┘
```

## This PR:

```
┌───────────────┬─────────────┬─────┬────────────┬────────────┬────────────┬────────────┬──────────┬────────┬─────────┐
│ Architectures │ Float_types │  Ns │        min │     median │       mean │        max │   memory │ allocs │ samples │
├───────────────┼─────────────┼─────┼────────────┼────────────┼────────────┼────────────┼──────────┼────────┼─────────┤
│           CPU │     Float32 │  32 │   8.592 ms │   8.835 ms │   9.210 ms │  11.790 ms │ 2.57 MiB │   7000 │      10 │
│           CPU │     Float32 │  64 │  14.176 ms │  16.614 ms │  16.460 ms │  19.553 ms │ 3.28 MiB │  12627 │      10 │
│           CPU │     Float32 │ 128 │  47.150 ms │  48.754 ms │  48.730 ms │  51.322 ms │ 3.57 MiB │  15635 │      10 │
│           CPU │     Float32 │ 256 │ 319.533 ms │ 326.286 ms │ 329.362 ms │ 347.420 ms │ 4.09 MiB │  36448 │      10 │
│           CPU │     Float64 │  32 │   8.080 ms │   8.429 ms │   8.621 ms │  10.488 ms │ 3.30 MiB │   7006 │      10 │
│           CPU │     Float64 │  64 │  13.719 ms │  14.112 ms │  15.023 ms │  18.725 ms │ 4.12 MiB │  12732 │      10 │
│           CPU │     Float64 │ 128 │  65.934 ms │  70.700 ms │  73.488 ms │  96.479 ms │ 4.49 MiB │  15963 │      10 │
│           CPU │     Float64 │ 256 │ 465.111 ms │ 476.674 ms │ 476.986 ms │ 488.592 ms │ 5.31 MiB │  56982 │      10 │
│           GPU │     Float32 │  32 │   2.081 ms │   2.199 ms │   2.346 ms │   3.654 ms │ 1.56 MiB │   4670 │      10 │
│           GPU │     Float32 │  64 │   2.405 ms │   2.469 ms │   2.611 ms │   3.793 ms │ 1.54 MiB │   4681 │      10 │
│           GPU │     Float32 │ 128 │   2.901 ms │   2.998 ms │   3.153 ms │   4.661 ms │ 1.54 MiB │   4681 │      10 │
│           GPU │     Float32 │ 256 │  16.744 ms │  19.228 ms │  18.980 ms │  19.308 ms │ 1.54 MiB │   4681 │      10 │
│           GPU │     Float64 │  32 │   2.099 ms │   2.365 ms │   2.466 ms │   3.583 ms │ 2.00 MiB │   4644 │      10 │
│           GPU │     Float64 │  64 │   2.525 ms │   2.717 ms │   3.028 ms │   4.566 ms │ 1.98 MiB │   4655 │      10 │
│           GPU │     Float64 │ 128 │   4.819 ms │   4.914 ms │   4.991 ms │   5.825 ms │ 1.98 MiB │   4655 │      10 │
│           GPU │     Float64 │ 256 │  24.680 ms │  36.323 ms │  35.144 ms │  36.496 ms │ 1.98 MiB │   4655 │      10 │
└───────────────┴─────────────┴─────┴────────────┴────────────┴────────────┴────────────┴──────────┴────────┴─────────┘
```